feat: edit and remove can only be perfomed by the story authors

+++ src/modules/Story/Remove/StoryRemoveService.ts
@@ -3,7 +3,7 @@ import { IStoryRepository } from '../../../repositories/Story/IStoryRepository';
 import { Messages, RequestError } from '../../../utils';
 
 export interface IStoryRemoveService {
-  handle(id: string): Promise<void>;
+  handle(id: string, authorId: string): Promise<void>;
 }
 
 class StoryRemoveService {
@@ -11,10 +11,14 @@ class StoryRemoveService {
     this.repository = repository;
   }
 
-  public handle = async (id: string) => {
+  public handle = async (id: string, authorId: string) => {
     const existingStory = await this.repository.findById(id);
     if (!existingStory) {
       throw new RequestError(Messages.STORY_NOT_FOUND, StatusCodes.NOT_FOUND);
+    }
+
+    if (existingStory.authorId !== authorId) {
+      throw new RequestError(Messages.YOU_DONT_HAVE_PERMISSION, StatusCodes.FORBIDDEN);
     }
 
     await this.repository.remove(id);
